Refs #26709 -- Checked allow_migrate_model() in Add/RemoveIndex operations.

+++ django/db/migrations/operations/models.py
@@ -772,11 +772,13 @@ class AddIndex(IndexOperation):
 
     def database_forwards(self, app_label, schema_editor, from_state, to_state):
         model = to_state.apps.get_model(app_label, self.model_name)
-        schema_editor.add_index(model, self.index)
+        if self.allow_migrate_model(schema_editor.connection.alias, model):
+            schema_editor.add_index(model, self.index)
 
     def database_backwards(self, app_label, schema_editor, from_state, to_state):
         model = from_state.apps.get_model(app_label, self.model_name)
-        schema_editor.remove_index(model, self.index)
+        if self.allow_migrate_model(schema_editor.connection.alias, model):
+            schema_editor.remove_index(model, self.index)
 
     def deconstruct(self):
         kwargs = {
@@ -812,15 +814,17 @@ class RemoveIndex(IndexOperation):
 
     def database_forwards(self, app_label, schema_editor, from_state, to_state):
         model = from_state.apps.get_model(app_label, self.model_name)
-        from_model_state = from_state.models[app_label, self.model_name_lower]
-        index = from_model_state.get_index_by_name(self.name)
-        schema_editor.remove_index(model, index)
+        if self.allow_migrate_model(schema_editor.connection.alias, model):
+            from_model_state = from_state.models[app_label, self.model_name_lower]
+            index = from_model_state.get_index_by_name(self.name)
+            schema_editor.remove_index(model, index)
 
     def database_backwards(self, app_label, schema_editor, from_state, to_state):
         model = to_state.apps.get_model(app_label, self.model_name)
-        to_model_state = to_state.models[app_label, self.model_name_lower]
-        index = to_model_state.get_index_by_name(self.name)
-        schema_editor.add_index(model, index)
+        if self.allow_migrate_model(schema_editor.connection.alias, model):
+            to_model_state = to_state.models[app_label, self.model_name_lower]
+            index = to_model_state.get_index_by_name(self.name)
+            schema_editor.add_index(model, index)
 
     def deconstruct(self):
         kwargs = {
